requirements.txt
requests
python-telegram-bot>=20.0a1
